Adds the possibility to load solps from balance.nc
Changes the required structure of the config dictionary:
        config['plasma']['edge']['mds_server']
        config['plasma']['edge']['mds_solps_reference']
        config['plasma']['edge']['solps_directory']
are no longer used.

The value of config['plasma']['edge']['mesh'] is not checked and the
mesh is always returned.

The config['plasma']['edge']['SOLPS_format'] now decided the way
config['plasma']['edge']["solps_reference"] will be interpreted:

1. if config['plasma']['edge']['SOLPS_format'] is 'MDSplus' then the
config['plasma']['edge']["solps_reference"] is expected to be a
dictionary with two keys "mds_server" and "reference":
        config["plasma"]["edge"]["solps_reference"]["mds_server"] is the
        reference to the mdsplus server.
        config["plasma"]["edge"]["solps_reference"]["reference"] is the
        reference to the SOLPS simulation.
2. if config['plasma']['edge']['SOLPS_format'] is 'RawFiles' then the
config['plasma']['edge']["solps_reference"] is expected to be the path
to the folder containing the solps raw files as fort.44.
3. if config['plasma']['edge']['SOLPS_format'] is 'balance.nc' then the
config['plasma']['edge']["solps_reference"] is expected to be the path
to the solps balance.nc file.